sent message details to backend using fetch and separated the code of contact form
